requirements.txt
asgiref==3.8.1
beautifulsoup4==4.9.3
boto3==1.20.41
botocore==1.23.41
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distlib==0.3.8
Django==3.2
django-appconf==1.0.5
django-background-tasks==1.2.5
django-bitcoin==0.2
django-ckeditor==5.8.0
django-ckeditor-updated==4.4.4
django-compat==1.0.15
django-crontab==0.7.1
django-filter==21.1
django-ipware==7.0.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-session-timeout==0.1.0
django-storages==1.12.3
djangorestframework==3.11.2
ecdsa==0.16.0
filelock==3.13.1
future==1.0.0
geocoder==1.38.1
gunicorn==20.1.0
humanize==4.9.0
idna==2.10
jmespath==0.10.0
mysql-client==0.0.1
mysql-connector-python==8.0.33
Pillow
platformdirs==4.1.0
protobuf==3.20.3
pyaes==1.6.1
PyMySQL==1.0.3
pyotp==2.9.0
python-dateutil==2.8.2
python-ipware==3.0.0
pytz==2021.1
ratelim==0.1.6
requests==2.31.0
s3transfer==0.5.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
typing_extensions==4.5.0
tzdata==2022.7
ua-parser==0.18.0
urllib3==1.26.19
uuid==1.30
virtualenv==20.25.0
whitenoise==5.3.0
